Changed IsDefault to IsInitial
These semantics more accurately represent the intention.  Whereas
"default" can be taken to mean the default memory allocation, the
default value for a type, etc. "initial" refers to the initial object in
a category and is therefore more consistent with the 'Source' metaphor
(as in source of a mapping/morphism)